Added preview for BottleMathModalView, Changed bottle color
