+++ MetaCocktailsSwiftData/Views/CocktailBatchingCalculator/CBCLoadedCocktailView.swift
@@ -196,56 +196,4 @@ struct SplitBatchNavigationButton: View {
     }
 }
 
-struct BottleMathModalView: View {
-    @EnvironmentObject var viewModel: CBCViewModel
-    @FocusState.Binding var isFocused: Bool
-    @State private var numberOfBottlesText: String = ""
-    @State private var bottlesSizeText: String = ""
-    
-    var body: some View {
-        VStack {
-            Text("How many bottles do you have?")
-                .font(FontFactory.formLabel18)
-            HStack {
-                Button {
-                    
-                } label: {
-                    Image(systemName: "plus.circle")
-                }
-                TextField("\(numberOfBottlesText)", text: $numberOfBottlesText)
-                    .font(FontFactory.formLabel18)
-                    .multilineTextAlignment(.center)
-                    .padding(.vertical, 5)
-                    .keyboardType(.numberPad)
-                    .frame(width: 80)
-                    .background(Color(UIColor.systemGray5))
-                    .cornerRadius(10)
-                    .focused($isFocused)
-                Button {
-                    
-                } label: {
-                    Image(systemName: "minus.circle")
-                }
-                
-            }
-            
-            HStack{
-                Text("What volume are the bottles?")
-                    .font(FontFactory.formLabel18)
-                TextField("\(numberOfBottlesText)", text: $numberOfBottlesText)
-                    .font(FontFactory.formLabel18)
-                    .multilineTextAlignment(.center)
-                    .padding(.vertical, 5)
-                    .keyboardType(.numberPad)
-                    .frame(width: 80)
-                    .background(Color(UIColor.systemGray5))
-                    .cornerRadius(10)
-                    .focused($isFocused)
-                Text("ml")
-                    .font(FontFactory.formLabel18)
-            }
-            
-        }
-        .jamesHeader("Calculate by number of available bottles:")
-    }
-}
+
